Fixed artifactId issue in pom and made other Maven/Ant improvements.
- The artifactId in the core project was 'sommer', but the other modules
depending on it were trying to use 'sommer-core': the core module is now
called 'sommer-core'.
- Added a root module (pom packaging) that refers to the other sub-modules:
it defines the dependencies versions consistently for the sub-modules.
- Changed groupId from 'net.java.rdf.sommer' to 'net.java.dev.sommer' for
consistency with domain name (required if/when going to central maven
repository). This is only for the groupId, the package names are unchanged.
- Added 'classpathref' (on top of 'classpath') attribute in the Ant task
to be able to use the classpath reference given my Maven directly.
- Added 'misc/SimpleDemonstration' module, can be run with
"mvn exec:exec -P run-demo"


git-svn-id: https://sommer.dev.java.net/svn/sommer/maven/sommer@459 a925da88-a912-0410-a1af-dd118eaf9e3a